Have the Engine store a global message delay generator
By having a global delay generator, the delay generator used for the
simulations is completely independent of the topology. As a result, the
message delay generator can be changed much more easily and the same
topology instance can be used for multiple simulations with different
message delay generators.

The modified Engine, Exporter, and RandomDelayGenerator classes were
placed inside the simulator2 package to avoid compilation errors.