Fix issue when disabling neighbor without route
Before this commit, if a neighbor was set as disabled in the routing
table and a route was never set for that neighbor, then the neighbor was
not disabled. After this commit, any neighbor can be set as disabled
regardless of whether or not a route was set for it before.

A new test was added for the routing table to ensure this behavior is
verified.