Adjust notifications to use new architecture
For now, the MessageReceived and MessageSent notifications have been
eliminated because they cause issues with the new architecture. These
notifications should be generic for any messages and they are notified
by a node which is generic. However, this means that there must be a
notifier instance for each message type. In other words, notifiers can
no longer be objects. This poses a problem that will be addressed in the
future.

The notification collectors are now part of the new `utils` package.
Although, they are currently used for tests these classes are utility
classes and therefore it makes more sense to put them in the `utils`
package instead of placing them in the `testing` package.