Re-add support for Message Sent and Received notifications
The new notifications now carry a generic message, which means that a
listener will have to downcast the route if it wants to access its
properties. However, most of the time when listening to this
notifications the actual message properties is not important but the
fact that a message was sent or received. This seems to be a good
solution, at least for now.