Add CLI application
Runners now take an application on each run call. This way, the runner
can keep the application informed of progress or eventual errors. This
design allows the application to handle all errors leaving to the runner
just the responsibility to setup the simulation environment and run the
executions.

The topology reader handlers are now implemented inside a sealed class.
This allows the use of the `when` statement to determine which reader is
being used.

Included a DummyApplication that can be used for testing.