Have basic data collector count disconnected nodes
At the end of a simulation the data collector counts the number of nodes
which the protocol did not select any valid route. To accomplish this,
the End Notification now carries the topology used for the simulation.
The Protocol interface also needed to be changed to include the selected
route property.