Add a default route to the advertisements
This modification will totally break the core's interface, which in turn
will break many classes in the 'simulation' package. However, this
enables setting some default route for a destination before advertising
it. This makes sense for BGP and maybe for other protocols that may be
implemented using the 'core' included here.