Introduce initializers in the existing code
The InputArgumentsParser is now only responsible for parsing and
validating the user inputs, delegating the initialization and loading
task to the initializer. Parsing input arguments now returns an
initializer initialized with the parameters specified by the user.

Runner and Execution now are generic types that require the type of
route to be specified.

There are still many FIXME and TODO notes that must be addressed.