Refactor metadata collection and printing
Metadata is now collected in multiple points of the code and only
printed to a file by the runner. This way, it is not necessary to
include extra information in the Runner just to include it in the
metadata.

A KeyValueWriter was added to write the metadata information. But, this
component can also be used for other purposes.