Change Engine's simulate() siganture to support advertisement
This broke almost every integration test. To simplify things, a function
called simulate() was added to the 'testing' package with the same
signature as the previous method. This function calls the Engine's
simulate() method with an advertisement at time 0 with a BGP self route.
All tests were than updated to use this function.