Fix BGP's advertise always exporting the default
Before, BGP's advertise() method exported the default route every time.
Now the default route is added to the protocol through the process()
method, which makes the route be only exported if it is selected.

Since the default route is now learned and selected, the corresponding
notifications are issued, which made some notification tests fail. They
were fixed.