Turn Scheduler into a class instead of an Object
Scheduler was an object to allow access to it without a reference being
passed all the time. But, meanwhile, that was accomplished by having the
Engine be an object and containing the scheduler used as a property.
Having this, classes are supposed to access the scheduler assigned to
the Engine object. Almost all classes were already doing that. Only the
Exporter class was, incorrectly, accessing the Scheduler object. With
this change, the Exporter class was also updated to access the scheduler
through the Engine object.